SecSplitLLM/baselines/fullsplitfed/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "splitfed"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.19.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "bhussn"

[tool.flwr.app.components]
serverapp = "splitfed.server.server_fl:app"
clientapp = "splitfed.clients.client_splitfed:app"

[tool.flwr.app.config]
num-server-rounds = 30
fraction-fit = 0.5
local-epochs = 2

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 15
